Add bookings

+++ src/main/java/ru/practicum/shareit/item/service/repository/impl/ItemRepositoryImpl.java
@@ -12,10 +12,10 @@ import java.util.stream.Stream;
 @Component
 public class ItemRepositoryImpl implements ItemRepository {
 
-    private static Map<Integer, Item> items = new HashMap<>();
+    private static Map<Long, Item> items = new HashMap<>();
 
     @Override
-    public Item get(int id) {
+    public Item get(long id) {
         return items.get(id);
     }
